+++ src/basicplus.js
@@ -1,4 +1,4 @@
-/** ****************************************************************************
+/** ************************************************************************
  *
  * A ...
  *
@@ -11,22 +11,27 @@
  *  . none,
  *
  *
+ * Private Static Methods:
+ *  . _setTestMode                returns internal objects for testing purpose,
+ *
+ *
  * Public Static Methods:
  *  . noConflict                  returns a reference to this ES6Pakket object,
+ *  . whoami                      returns the library name and version,
  *  . getString                   returns a string,
  *  . getArray                    returns an array,
  *
  *
  *
- * @namespace    ES6Pakket
+ * @namespace    -
  * @dependencies none
  * @exports      -
  * @author       -
  * @since        0.0.0
  * @version      -
- * ************************************************************************** */
-/* global root, */
-/* eslint-disable one-var, semi-style */
+ * ********************************************************************** */
+/* global root */
+/* eslint-disable one-var, semi-style, no-underscore-dangle */
 
 
 // -- Vendor Modules
@@ -46,15 +51,41 @@ const previousES6Pakket = root.ES6Pakket;
 // -- Local Variables
 
 
-// -- Public Static Methods ------------------------------------------------
+// -- Main -----------------------------------------------------------------
 
 const ES6Pakket = {
 
+
+  // Useful to retrieve the library name and version when it is
+  // embedded in another library as an object:
+  _library: { name: '{{lib:name}}', version: '{{lib:version}}' },
+
+
+  // -- Private Static Methods ---------------------------------------------
+
+  /**
+   * Returns the internal objects for testing purpose.
+   * (must not be deleted)
+   *
+   * @method ()
+   * @private
+   * @param {}                -,
+   * @returns {Object}        returns a list of internal objects,
+   * @since 0.0.0
+   */
+  _setTestMode() {
+    return [];
+  },
+
+
+  // -- Public Static Methods ----------------------------------------------
+
   /**
    * Returns a reference to this ES6Pakket object.
+   * (must not be deleted)
    *
    * Nota:
-   * Running ES6Pakket in noConflic mode, returns the ES6Pakket variable to
+   * Running ES6Pakket in noConflict mode, returns the ES6Pakket variable to
    * its previous owner.
    *
    * @function ()
@@ -63,15 +94,29 @@ const ES6Pakket = {
    * @returns {Object}      returns the ES6Pakket object,
    * @since 0.0.0
    */
-  /* istanbul ignore next */
   noConflict() {
     /* eslint-disable-next-line no-param-reassign */
     root.ES6Pakket = previousES6Pakket;
     return this;
   },
+
+  /**
+   * Returns the library name and version.
+   * (must not be deleted)
+   *
+   * @method ()
+   * @public
+   * @param {}              -,
+   * @returns {Object}      returns the library name and version,
+   * @since 0.0.0
+   */
+  whoami() {
+    return this._library;
+  },
 };
 
-// Attaches a constant to ES6Pakket that provides the version of the lib.
+// Attaches constants to ES6Pakket that provide name and version of the lib.
+ES6Pakket.NAME = '{{lib:name}}';
 ES6Pakket.VERSION = '{{lib:version}}';
 
 
@@ -109,4 +154,4 @@ extend(ES6Pakket, {
 // -- Export
 export default ES6Pakket;
 
-/* eslint-enable one-var, semi-style */
+/* eslint-enable one-var, semi-style, no-underscore-dangle */