+++ src/functional-shared.js
@@ -1,4 +1,4 @@
-/** ****************************************************************************
+/** ************************************************************************
  *
  * A ...
  *
@@ -11,7 +11,11 @@
  *
  *
  * Constructor:
- *  . ES6Pakket                  creates and returns the ES6Pakket object,
+ *  . ES6Pakket                   creates and returns the ES6Pakket object,
+ *
+ *
+ * Private Static Methods:
+ *  . _setTestMode                returns internal objects for testing purpose,
  *
  *
  * Public Static Methods:
@@ -19,20 +23,21 @@
  *
  *
  * Public Methods:
+ *  . whoami                      returns the library name and version,
  *  . getString                   returns a string,
  *  . getArray                    returns an array,
  *
  *
  *
- * @namespace    ES6Pakket
+ * @namespace    -
  * @dependencies none
  * @exports      -
  * @author       -
  * @since        0.0.0
  * @version      -
- * ************************************************************************** */
+ * ********************************************************************** */
 /* global root */
-/* eslint-disable one-var, semi-style */
+/* eslint-disable one-var, semi-style, no-underscore-dangle */
 
 
 // -- Vendor Modules
@@ -55,7 +60,7 @@ let methods
   ;
 
 
-// -- Public -------------------------------------------------------------------
+// -- Public ---------------------------------------------------------------
 
 /**
  * Returns the ES6Pakket object.
@@ -69,22 +74,45 @@ let methods
  */
 const ES6Pakket = function(name) {
   const obj = {};
+  obj._library = {
+    name: '{{lib:name}}',
+    version: '{{lib:version}}',
+  };
   obj.name = name;
   extend(obj, methods);
   return obj;
 };
 
-// Attaches a constant to ES6Pakket that provides the version of the lib.
+// Attaches constants to ES6Pakket that provide name and version of the lib.
+ES6Pakket.NAME = '{{lib:name}}';
 ES6Pakket.VERSION = '{{lib:version}}';
 
 
-// -- Public Static Methods ----------------------------------------------------
+// -- Private Static Methods -----------------------------------------------
+
+/**
+ * Returns the internal objects for testing purpose.
+ * (must not be deleted)
+ *
+ * @method ()
+ * @private
+ * @param {}                -,
+ * @returns {Object}        returns a list of internal objects,
+ * @since 0.0.0
+ */
+ES6Pakket._setTestMode = function() {
+  return [];
+};
+
+
+// -- Public Static Methods ------------------------------------------------
 
 /**
  * Returns a reference to this ES6Pakket object.
+ * (must not be deleted)
  *
  * Nota:
- * Running ES6Pakket in noConflic mode, returns the ES6Pakket variable to
+ * Running ES6Pakket in noConflict mode, returns the ES6Pakket variable to
  * its previous owner.
  *
  * @method ()
@@ -93,7 +121,6 @@ ES6Pakket.VERSION = '{{lib:version}}';
  * @returns {Object}        returns the ES6Pakket object,
  * @since 0.0.0
  */
-/* istanbul ignore next */
 ES6Pakket.noConflict = function() {
   /* eslint-disable-next-line no-param-reassign */
   root.ES6Pakket = previousES6Pakket;
@@ -101,9 +128,23 @@ ES6Pakket.noConflict = function() {
 };
 
 
-// -- Public Methods -----------------------------------------------------------
+// -- Public Methods -------------------------------------------------------
 
 methods = {
+
+  /**
+   * Returns the library name and version.
+   * (must not be deleted)
+   *
+   * @method ()
+   * @public
+   * @param {}              -,
+   * @returns {Object}      returns the library name and version,
+   * @since 0.0.0
+   */
+  whoami() {
+    return this._library;
+  },
 
   /**
    * Returns a string.
@@ -136,4 +177,4 @@ methods = {
 // -- Export
 export default ES6Pakket;
 
-/* eslint-enable one-var, semi-style */
+/* eslint-enable one-var, semi-style, no-underscore-dangle */