+++ test/main.js
@@ -1,70 +1,38 @@
 // ESLint declarations:
-/* global describe, it */
-/* eslint one-var: 0, no-unused-vars: 0, semi-style: 0 */
+/* global describe */
+/* eslint one-var: 0, semi-style: 0 */
 
 
 // -- Vendor Modules
-const should     = require('chai').should()
-    , { expect } = require('chai')
-    ;
 
 
 // -- Local Modules
 const ES6Pakket = require('../index')
+    // ES6Pakket = require('../src/prototypal').default
+    , pack      = require('../package.json')
+    , testlib   = require('./int/libprototypal')
     ;
 
 
 // -- Local Constants
+const libname = 'ES6Pakket';
 
 
 // -- Local Variables
 
 
 // -- Main
+
+// Nota:
+// If you choose 'ES6Pakket = require('../index')', 'display-coverage' will
+// show the coverage of all the library in one file.
+//
+// If you want to display the coverage file by file, you must choose
+// 'ES6Pakket = require('../src/prototypal').default'. But, in this case,
+// the build isn't done, so you should pass '{{lib:name}}' as libname and
+// '{{lib:version}}' as the library version.
+
 describe('Test ES6Pakket:', () => {
-  // Test the lib:
-  describe('Test ES6Pakket.VERSION and ES6Pakket.noConflict:', () => {
-    it('Expects ES6Pakket.VERSION to return a string.', () => {
-      expect(ES6Pakket.VERSION).to.be.a('string');
-    });
-    it('Expects ES6Pakket.noConflict to return a function.', () => {
-      expect(ES6Pakket.noConflict).to.be.a('function');
-    });
-  });
-
-  describe('Test ES6Pakket constructor and methods:', () => {
-    const o = ES6Pakket();
-
-    it('Expects ES6Pakket() to return an object.', () => {
-      expect(o).to.be.an('object');
-    });
-
-    it('Expects this object to own the property "getString" that is a function.', () => {
-      expect(o).to.have.property('getString').that.is.a('function');
-    });
-
-    it('Expects this object to own the property "getArray" that is a function.', () => {
-      expect(o).to.have.property('getArray').that.is.a('function');
-    });
-
-    it('Expects "getString" to return the string "I am a string!".', () => {
-      expect(o.getString()).to.be.a('string').that.is.equal('I am a string!');
-    });
-
-    it('Expects "getArray" to return an array with 3 elements.', () => {
-      expect(o.getArray()).to.be.an('array').that.has.lengthOf(3);
-    });
-
-    it('Expects the first element to be equal to 1.', () => {
-      expect(o.getArray()[0]).to.be.a('number').that.is.equal(1);
-    });
-
-    it('Expects the second element to be equal to 2.', () => {
-      expect(o.getArray()[1]).to.be.a('number').that.is.equal(2);
-    });
-
-    it('Expects the third element to be equal to 3.', () => {
-      expect(o.getArray()[2]).to.be.a('number').that.is.equal(3);
-    });
-  });
+  // testlib(ES6Pakket, '{{lib:name}}', '{{lib:version}}');
+  testlib(ES6Pakket, libname, pack.version);
 });