+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### HEAD
 
-### 0.0.3 (August 8, 2020)
+### 0.0.3 (August 18, 2020)
 
   * Fixed a few typos,
   * Added the ESM module to enable Mocha testing from the source files,
@@ -13,6 +13,8 @@
   * Fourth release candidate,
   * Updated the project dependencies,
   * Fifth release candidate,
+  * Added the method whoami,
+  * Reorganised the test folder,
   * ...,